static_site_builder/requirements.txt
fastapi>=0.115.6
uvicorn>=0.34.0
jinja2>=3.1.4
python-multipart>=0.0.9
crewai>=0.86.0
chromadb>=0.5.23
openai>=1.58.1
pydantic>=2.9.2
python-dotenv>=1.0.1